ocr_ai_backend/pyproject.toml
[project]
name = "ocr-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn[standard]>=0.30.0",
	"python-multipart>=0.0.9",
	"pillow>=10.0.0",
	"easyocr>=1.7.1",
	# EasyOCR relies on torch/torchvision; these wheels are platform-specific.
	# If installation fails, install appropriate torch/torchvision versions manually.
	"torch>=2.1.0; platform_machine!='i386'",
	"torchvision>=0.16.0; platform_machine!='i386'",
	# VLM captioning (optional, lazy-loaded)
	"transformers>=4.44.0",
	"huggingface-hub>=0.23.0",
	# Online VLM & env loading
	"httpx>=0.27.0",
	"python-dotenv>=1.0.1",
	"numpy>=1.26.0",
]

[project.scripts]
ocr-ai-backend = "main:run"

[tool.uv]
package = true
